refactor: split UI API contracts
